[MOB-15975] -- Rid of hardcoded names of plugins in bash scripts

+++ buildSrc/src/main/kotlin/ru/hh/plugins/gradle/build_all_plugins/BuildAllPluginsTask.kt
@@ -27,11 +27,12 @@ open class BuildAllPluginsTask @Inject constructor(
     fun buildAllPlugins() {
         inputFiles.forEach { pluginArchive ->
             val newPluginArchiveDestinationFile = outputFiles.get().asFile.resolve(pluginArchive.name)
-
-            Files.move(
-                pluginArchive.toPath(),
-                newPluginArchiveDestinationFile.toPath()
-            )
+            if (newPluginArchiveDestinationFile.exists().not()) {
+                Files.move(
+                    pluginArchive.toPath(),
+                    newPluginArchiveDestinationFile.toPath()
+                )
+            }
         }
     }
